# RealworldWeb/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "RealWorld Conduit API",
    "description" : "Conduit API documentation",
    "contact" : {
      "name" : "RealWorld",
      "url" : "https://realworld.how"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://api.realworld.io/api"
  } ],
...
  "paths" : {
...
    "/profiles/{username}" : {
      "get" : {
        "tags" : [ "Profile" ],
        "summary" : "Get a profile",
        "description" : "Get a profile of a user of the system. Auth is optional",
        "operationId" : "GetProfileByUsername",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "Username of the profile to get",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Profile",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_1"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        }
      }
    },
    "/profiles/{username}/follow" : {
      "post" : {
        "tags" : [ "Profile" ],
        "summary" : "Follow a user",
        "description" : "Follow a user by username",
        "operationId" : "FollowUserByUsername",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "Username of the profile you want to follow",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Profile",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_1"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Profile" ],
        "summary" : "Unfollow a user",
        "description" : "Unfollow a user by username",
        "operationId" : "UnfollowUserByUsername",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "Username of the profile you want to unfollow",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Profile",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_1"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ]
      }
    },
...
    "/articles/{slug}/comments" : {
      "get" : {
        "tags" : [ "Comments" ],
        "summary" : "Get comments for an article",
        "description" : "Get the comments for an article. Auth is optional",
        "operationId" : "GetArticleComments",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "description" : "Slug of the article that you want to get comments for",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Multiple comments",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_3"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Comments" ],
        "summary" : "Create a comment for an article",
        "description" : "Create a comment for an article. Auth is required",
        "operationId" : "CreateArticleComment",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "description" : "Slug of the article that you want to create a comment for",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/NewCommentRequest"
        },
        "responses" : {
          "200" : {
            "description" : "Single comment",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "x-codegen-request-body-name" : "comment"
      }
    },
    "/articles/{slug}/comments/{id}" : {
      "delete" : {
        "tags" : [ "Comments" ],
        "summary" : "Delete a comment for an article",
        "description" : "Delete a comment for an article. Auth is required",
        "operationId" : "DeleteArticleComment",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "description" : "Slug of the article that you want to delete a comment for",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the comment you want to delete",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "No content",
            "content" : { }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Profile" : {
        "required" : [ "bio", "following", "image", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          },
          "following" : {
            "type" : "boolean"
          }
        },
        "example" : {
          "image" : "image",
          "following" : true,
          "bio" : "bio",
          "username" : "username"
        }
      },
...
      "Comment" : {
        "required" : [ "author", "body", "createdAt", "id", "updatedAt" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "body" : {
            "type" : "string"
          },
          "author" : {
            "$ref" : "#/components/schemas/Profile"
          }
        },
        "example" : {
          "createdAt" : "2000-01-23T04:56:07.000+00:00",
          "author" : {
            "image" : "image",
            "following" : true,
            "bio" : "bio",
            "username" : "username"
          },
          "id" : 0,
          "body" : "body",
          "updatedAt" : "2000-01-23T04:56:07.000+00:00"
        }
      },
      "NewComment" : {
        "required" : [ "body" ],
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string"
          }
        }
      },
      "GenericErrorModel" : {
        "required" : [ "errors" ],
        "type" : "object",
        "properties" : {
          "errors" : {
            "$ref" : "#/components/schemas/GenericErrorModel_errors"
          }
        }
      },
...
      "inline_response_200_1" : {
        "required" : [ "profile" ],
        "type" : "object",
        "properties" : {
          "profile" : {
            "$ref" : "#/components/schemas/Profile"
          }
        },
        "example" : {
          "profile" : {
            "image" : "image",
            "following" : true,
            "bio" : "bio",
            "username" : "username"
          }
        }
      },
...
      "inline_response_200_3" : {
        "required" : [ "comments" ],
        "type" : "object",
        "properties" : {
          "comments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Comment"
            }
          }
        },
        "example" : {
          "comments" : [ {
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "id" : 0,
            "body" : "body",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00"
          }, {
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "id" : 0,
            "body" : "body",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00"
          } ]
        }
      },
      "inline_response_200_4" : {
        "required" : [ "comment" ],
        "type" : "object",
        "properties" : {
          "comment" : {
            "$ref" : "#/components/schemas/Comment"
          }
        },
        "example" : {
          "comment" : {
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "id" : 0,
            "body" : "body",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00"
          }
        }
      },
...
      "GenericErrorModel_errors" : {
        "required" : [ "body" ],
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "requestBodies" : {
...
      "NewCommentRequest" : {
        "description" : "Comment you want to create",
        "content" : {
          "application/json" : {
            "schema" : {
              "required" : [ "comment" ],
              "type" : "object",
              "properties" : {
                "comment" : {
                  "$ref" : "#/components/schemas/NewComment"
                }
              }
            }
          }
        },
        "required" : true
      }
    },
    "securitySchemes" : {
      "Token" : {
        "type" : "apiKey",
        "description" : "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "name" : "Authorization",
        "in" : "header"
      }
    }
  }
}